refactor(#13): streamline platform abstraction and event_flag interface
- Move event_flag implementations to header files as static inline functions in C11 and Pthread ports.

- Convert event_flag API in .in templates to use static inline prototypes.

- Improve task management in C11 port by implementing a local current task tracker with sanity checks, replacing pthread_getname_np.

- Enhance documentation for tasks and thread-safe queues across all platforms using Doxygen style.

- Update CMakeLists.txt to remove event_flag.c from build sources and add _GNU_SOURCE definition.

- Refine mutex and ts_queue header templates with detailed API documentation.
