+++ CMakeLists.txt
@@ -93,7 +93,6 @@ if(GIT_FOUND
   endif()
 endif()
 
-# Require out-of-source builds
 file(TO_CMAKE_PATH "${CMAKE_CURRENT_BINARY_DIR}/CMakeLists.txt" LOC_PATH)
 if(EXISTS "${LOC_PATH}")
   message(FATAL_ERROR "You cannot build in a source directory (or any directory with a CMakeLists.txt file). "
