+++ inc/cutils/freertos/task.h
@@ -32,6 +32,10 @@
 extern "C" {
 #endif
 
+/** @name Task Priority Levels
+ *  These macros define the priority range for tasks on this port.
+ *  @note In FreeRTOS, higher numbers typically represent higher priorities.
+ *  @{ */
 #define CUTILS_TASK_PRIORITY_LOWEST (0)
 #define CUTILS_TASK_PRIORITY_HIGHEST                                                               \
   (configMAX_PRIORITIES - 2) // Timer thread takes configMAX_PRIORITIES - 1
@@ -41,6 +45,7 @@ extern "C" {
   (CUTILS_TASK_PRIORITY_MEDIUM - ((CUTILS_TASK_PRIORITY_HIGHEST - CUTILS_TASK_PRIORITY_LOWEST) / 4))
 #define CUTILS_TASK_PRIORITY_MID_HIGH                                                              \
   (CUTILS_TASK_PRIORITY_MEDIUM + ((CUTILS_TASK_PRIORITY_HIGHEST - CUTILS_TASK_PRIORITY_LOWEST) / 4))
+/** @} */
 
 #define CUTILS_TASK_STACK_MIN_SIZE (configMINIMAL_STACK_SIZE)
 #define DEFAULT_TASK_PRIORITY      (CUTILS_TASK_PRIORITY_MEDIUM)
@@ -50,8 +55,14 @@ extern "C" {
 #define CUTILS_TASK_STACK_ALIGN portBYTE_ALIGNMENT
 #endif
 
+/**
+ * @brief Function signature for task entry points.
+ */
 typedef void (*task_func_t)(void *);
 
+/**
+ * @brief The internal representation of a task on the FreeRTOS port.
+ */
 typedef struct _task_t {
   TaskHandle_t task;
   StaticTask_t tcb;
@@ -60,6 +71,9 @@ typedef struct _task_t {
   void *ctx;
 } task_t;
 
+/**
+ * @brief Parameters used to initialize a task.
+ */
 typedef struct _task_create_params_t {
   task_t *task;
   StaticTask_t *tcb;
@@ -71,19 +85,29 @@ typedef struct _task_create_params_t {
   size_t stack_size;
 } task_create_params_t;
 
+/**
+ * @name Static Task Storage Macros
+ *  These macros provide a mechanism for statically allocating tasks and their stacks.
+ *  @{ */
 #define TASK_STATIC_STORE_T(name) task_static_store_##name##_t
 
-#define TASK_STATIC_STORE(name)   _task_store_##name
+/** @brief The symbol name for the static store of a given task name. */
+#define TASK_STATIC_STORE(name) _task_store_##name
 
+/**
+ * @brief Declares a structure that holds both the stack and the task control block.
+ */
 #define TASK_STATIC_STORE_DECL(name, stack_size)                                                   \
   typedef struct {                                                                                 \
     alignas(CUTILS_TASK_STACK_ALIGN) uint8_t stack[(stack_size)];                                  \
     task_t tsk;                                                                                    \
   } TASK_STATIC_STORE_T(name)
 
+/** @brief Defines the static store as a used symbol. */
 #define TASK_STATIC_STORE_DEF(name)                                                                \
   TASK_STATIC_STORE_T(name) TASK_STATIC_STORE(name) __attribute__((used))
 
+/** @brief Internal helper to map a static store to task create parameters. */
 #define TASK_INIT_CREATE_PARAMS_FROM_STORE(params, store_ptr, lbl, pri, fn, context)               \
   memset(&(params), 0, sizeof((params)));                                                          \
   (params).task = &(store_ptr)->tsk;                                                               \
@@ -95,20 +119,10 @@ typedef struct _task_create_params_t {
   (params).stack = (store_ptr)->stack;                                                             \
   (params).stack_size = sizeof((store_ptr)->stack)
 
+/** @brief Public macro to initialize create parameters from a named static store. */
 #define TASK_STATIC_INIT_CREATE_PARAMS(params, name, lbl, pri, fn, context)                        \
   TASK_INIT_CREATE_PARAMS_FROM_STORE(params, &TASK_STATIC_STORE(name), lbl, pri, fn, context)
-
-task_t *task_new_static(task_create_params_t *create_params);
-
-bool task_start(task_t *task);
-
-void task_destroy_static(task_t *task);
-
-uint64_t task_get_ticks(void);
-
-void task_sleep(uint32_t ms);
-
-void task_get_current_name(char *name, size_t string_length);
+/** @} */
 
 #ifdef __cplusplus
 }
