+++ src/pthread/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(Threads REQUIRED)
 
 file(GLOB PLATFORM_HEADER_LIST CONFIGURE_DEPENDS "${cutils_SOURCE_DIR}/inc/cutils/pthread/*.h")
 
-set(PLATFORM_SOURCES task.c event_flag.c)
+set(PLATFORM_SOURCES task.c)
 add_library(platform_abstraction STATIC "${PLATFORM_HEADER_LIST}" "${HEADER_LIST}"
                                         "${PLATFORM_SOURCES}")
 set_target_properties(platform_abstraction PROPERTIES LINKER_LANGUAGE C)
